thrasher-cover-similarity/pyproject.toml
[project]
name = "thrasher-covers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.14"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "numpy<2",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "scikit-learn>=1.6.1",
    "torch>=2.2",
    "torchvision>=0.17",
    "tqdm>=4.67.1",
    "transformers>=4.40.0",
    "umap-learn>=0.5.9.post2",
]
